+++ packages/stf/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @fumari/stf
 
+## 1.0.2
+
+### Patch Changes
+
+- bd8c33a: fix array access
+
+## 1.0.1
+
+### Patch Changes
+
+- 21e60c3: support cleaning namespace information
+
 ## 1.0.0
 
 ### Major Changes