+++ packages/story/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @fumadocs/story
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [38bd784]
+- Updated dependencies [bd8c33a]
+  - fumadocs-ui@16.6.6
+  - @fumari/stf@1.0.2
+
+## 0.0.9
+
+### Patch Changes
+
+- 21e60c3: support cleaning namespace information
+- Updated dependencies [21e60c3]
+  - @fumari/stf@1.0.1
+
 ## 0.0.8
 
 ### Patch Changes